+++ tests/cli/test_cli_admin.py
@@ -12,7 +12,7 @@ import os
 
 from click.testing import CliRunner
 
-from flowserv.cli.admin import cli
+from flowserv.client.cli.admin import cli
 from flowserv.config.api import FLOWSERV_API_BASEDIR
 from flowserv.config.database import FLOWSERV_DB
 from flowserv.model.database import TEST_URL