+++ tests/cli/test_cli_workflow.py
@@ -10,7 +10,7 @@
 
 import os
 
-from flowserv.cli.admin import cli
+from flowserv.client.cli.admin import cli
 
 
 DIR = os.path.dirname(os.path.realpath(__file__))